chore(plugin-doc-coverage): rename processer to processor
