+++ packages/plugin-doc-coverage/src/lib/runner/doc-processor.ts
