+++ packages/plugin-doc-coverage/tsconfig.lib.json
@@ -5,7 +5,7 @@
     "declaration": true,
     "types": ["node"]
   },
-  "include": ["src/**/*.ts", "src/lib/runner/doc-processer.js"],
+  "include": ["src/**/*.ts"],
   "exclude": [
     "vite.config.unit.ts",
     "vite.config.integration.ts",
